Implementing a custom MVCFactory
The amount of boilerplate and complexity is insane.

What is worse is that the custom service provider will
need to be maintained with every single minor and patch
release of Joomla as the original is considered INTERNALS
and changing it is supposedly not a b/c break.

Therefore the component's DI Container is impractical if
not useless. Derp.